+++ client/src/stores/index.js
@@ -1,0 +1,4 @@
+export * from './authStore';
+export * from './clipboardStore';
+export * from './presenceStore';
+export * from './adminStore';